Added build tag so that mandatory rule-id will be checked when `coraza.rule.mandatory_rule_id_check` build tag is defined.
