+++ internal/corazawaf/rule_mandatory_idcheck.go
@@ -1,0 +1,6 @@
+//go:build coraza.rule.mandatory_rule_id_check
+// +build coraza.rule.mandatory_rule_id_check
+
+package corazawaf
+
+var shouldDoMandatoryRuleIdCheck = true
