+++ internal/corazawaf/rulegroup_test.go
@@ -85,34 +85,3 @@ func TestRuleGroupDeleteByID(t *testing.T) {
 		t.Fatal("Unexpected remaining rule in the rulegroup")
 	}
 }
-
-func TestRuleIDMandatoryCondition(t *testing.T) {
-	r := NewRule()
-	r.Msg, _ = macro.NewMacro("test-rule")
-	r.Tags_ = []string{
-		"test/no-id",
-	}
-
-	rg := NewRuleGroup()
-	if err := rg.Add(r); err == nil {
-		t.Error("Expected error - rule without id should not be allowed")
-	}
-}
-
-func TestRuleIDDuplicate(t *testing.T) {
-	r := newTestRule(1)
-
-	rg := NewRuleGroup()
-	if err := rg.Add(r); err != nil {
-		t.Error("Failed to add rule to rulegroup")
-	}
-
-	if rg.Count() != 1 {
-		t.Error("Failed to add rule to rulegroup")
-	}
-
-	sr := newTestRule(1)
-	if err := rg.Add(sr); err == nil {
-		t.Error("Expected error - duplicate rule id should not be allowed")
-	}
-}
